+++ src/oemof/network/graph.py
@@ -46,58 +46,19 @@ def create_nx_graph(
     remove_edges: list of string tuples
         Edges to be removed e.g. [('resource_gas', 'gas_balance')]
 
-    Examples
-    --------
-    >>> import os
-    >>> import pandas as pd
-    >>> from oemof.network.network import Node
-    >>> from oemof.network.energy_system import EnergySystem
-    >>> import oemof.network.graph as graph
-    >>> datetimeindex = pd.date_range('1/1/2017', periods=3, freq='h')
-    >>> es = EnergySystem(timeindex=datetimeindex)
-    >>> b_gas = Node(label='b_gas')
-    >>> bel1 = Node(label='bel1')
-    >>> bel2 = Node(label='bel2')
-    >>> demand_el = Node(label='demand_el', inputs = [bel1])
-    >>> pp_gas = Node(label=('pp', 'gas'),
-    ...               inputs=[b_gas],
-    ...               outputs=[bel1])
-    >>> line_to2 = Node(label='line_to2', inputs=[bel1], outputs=[bel2])
-    >>> line_from2 = Node(label='line_from2',
-    ...                   inputs=[bel2], outputs=[bel1])
-    >>> es.add(b_gas, bel1, demand_el, pp_gas, bel2, line_to2, line_from2)
-    >>> my_graph = graph.create_nx_graph(es)
-    >>> # export graph as .graphml for programs like Yed where it can be
-    >>> # sorted and customized. this is especially helpful for large graphs
-    >>> # grph.create_nx_graph(es, filename="my_graph.graphml")
-    >>> [my_graph.has_node(nd)
-    ...  for nd in ['b_gas', 'bel1', "('pp', 'gas')", 'demand_el', 'tester']]
-    [True, True, True, True, False]
-    >>> list(nx.attracting_components(my_graph))
-    [{'demand_el'}]
-    >>> sorted(list(nx.strongly_connected_components(my_graph))[1])
-    ['bel1', 'bel2', 'line_from2', 'line_to2']
-    >>> new_graph = graph.create_nx_graph(energy_system=es,
-    ...                                  remove_nodes_with_substrings=['b_'],
-    ...                                  remove_nodes=["('pp', 'gas')"],
-    ...                                  remove_edges=[('bel2', 'line_from2')],
-    ...                                  filename='test_graph')
-    >>> [new_graph.has_node(nd)
-    ...  for nd in ['b_gas', 'bel1', "('pp', 'gas')", 'demand_el', 'tester']]
-    [False, True, False, True, False]
-    >>> my_graph.has_edge("('pp', 'gas')", 'bel1')
-    True
-    >>> new_graph.has_edge('bel2', 'line_from2')
-    False
-    >>> os.remove('test_graph.graphml')
-
     Notes
     -----
     Needs graphviz and networkx (>= v.1.11) to work properly.
     Tested on Ubuntu 16.04 x64 and solydxk (debian 9).
     """
+    warnings.warn(
+        "The function 'create_nx_graph' is deprecated,"
+        + " please use 'EnergySystem.to_nx_graph()' instead.",
+        category=FutureWarning,
+    )
+
     with warnings.catch_warnings():
-        # suppress ExperimentalFeatureWarnungs
+        # suppress ExperimentalFeatureWarning
         warnings.simplefilter("ignore")
 
         # construct graph from nodes and flows