+++ tests/test_energy_system.py
@@ -16,13 +16,15 @@ SPDX-License-Identifier: MIT
 
 from pathlib import Path
 
+import networkx as nx
 import pytest
-from networkx import DiGraph
+from oemof.tools.debugging import ExperimentalFeatureWarning
 
 from oemof.network import Edge
+from oemof.network import EnergySystem
 from oemof.network import Node
 from oemof.network import graph
-from oemof.network.energy_system import EnergySystem
+from oemof.network.network.nodes import QualifiedLabel
 
 
 def test_ensys_init():
@@ -30,9 +32,45 @@ def test_ensys_init():
     ensys = EnergySystem(nodes=[node])
     assert node in ensys.nodes
 
-    with pytest.warns(FutureWarning):
+    with pytest.warns(FutureWarning, match="entities"):
         ensys = EnergySystem(entities=[node])
         assert node in ensys.nodes
+
+    for attr in ["results", "timeindex", "timeincrement", "temporal"]:
+        with pytest.warns(FutureWarning, match=attr):
+            EnergySystem(**{attr: 1})
+
+
+def test_duplicate_label():
+    es = EnergySystem()
+    my_label1 = "test_01"
+    my_label2 = "test_02"
+    es.add(Node(label=my_label1))
+    es.add(Node(label=my_label2))
+    msg = (
+        r"EnergySystem already contains Node\(s\) with the following string"
+        + r' representation: "test_01", "test_02"'
+    )
+    with pytest.raises(ValueError, match=msg):
+        es.add(Node(label=my_label1), Node(label=my_label2))
+
+
+def test_duplicate_qualified_label():
+    es = EnergySystem()
+    my_label1 = QualifiedLabel(
+        ("test_01",),
+    )
+    my_label2 = QualifiedLabel(
+        ("test_02",),
+    )
+    es.add(Node(label=my_label1))
+    es.add(Node(label=my_label2))
+    msg = (
+        r"EnergySystem already contains Node\(s\) with the following string"
+        + r" representation: \"\('test_01',\)\", \"\('test_02',\)\""
+    )
+    with pytest.raises(ValueError, match=msg):
+        es.add(Node(label=my_label1), Node(label=my_label2))
 
 
 class TestDumpRestore:
@@ -263,6 +301,7 @@ class TestsEnergySystem:
         node1 = Node(label="node1")
         node2 = Node(label="node2")
         node3 = Node(label="node3", inputs={node2: Edge()})
+        node4 = Node(label="node4")
         self.es.check()  # empty, no problem
 
         self.es.add(node0)
@@ -284,10 +323,23 @@ class TestsEnergySystem:
         self.es.check()  # graph still consistent
 
         self.es.add(node2)
-        with pytest.raises(RuntimeError, match="not part of EnergySystem"):
+        with pytest.raises(
+            RuntimeError,
+            match="node3 not part of EnergySystem",
+        ):
             self.es.check()  # if node 2 is present, node3 also needs to be
 
         self.es.add(node3)
+        self.es.check()  # Now, everything is fine.
+
+        node3.inputs[node4] = Edge()
+        with pytest.raises(
+            RuntimeError,
+            match="node4 not part of EnergySystem",
+        ):
+            self.es.check()  # node 3 has a flow from node 4
+
+        self.es.add(node4)
         self.es.check()  # Now, everything is fine.
 
     def test_that_node_additions_are_signalled(self):
@@ -311,18 +363,138 @@ class TestsEnergySystem:
             "Probable reason: `subscriber` didn't get called."
         ).format(subscriber.called)
 
-    def test_graph(self):
+    def test_graph_function(self):
         fpath = Path(Path.home(), "test_graph_x345_efhu73.graphml")
-        my_graph = graph.create_nx_graph(self.es)
-        assert isinstance(my_graph, DiGraph)
+        with pytest.warns(
+            FutureWarning,
+            match="use 'EnergySystem.to_nx_graph\\(\\)' instead.",
+        ):
+            my_graph = graph.create_nx_graph(self.es)
+        assert isinstance(my_graph, nx.DiGraph)
 
         # make sure that test does not pass because of pre-existing file
         assert not fpath.is_file()
 
         # create graph file
-        my_graph = graph.create_nx_graph(self.es, filename=fpath)
-        assert isinstance(my_graph, DiGraph)
+        with pytest.warns(
+            FutureWarning,
+            match="use 'EnergySystem.to_nx_graph\\(\\)' instead.",
+        ):
+            my_graph = graph.create_nx_graph(self.es, filename=fpath)
+        assert isinstance(my_graph, nx.DiGraph)
         assert fpath.is_file()
 
         # clean up (delete graph file)
         fpath.unlink()
+
+    def test_graph_export(self):
+        node0 = Node(label="node 0")
+        node1 = Node(label="node 1", inputs={node0: Edge()})
+        self.es.add(node0, node1)
+
+        my_graph = self.es.to_networkx()
+        assert isinstance(my_graph, nx.DiGraph)
+        assert node0 in my_graph.nodes
+        assert node1 in my_graph.nodes
+
+        assert (node0, node1) in my_graph.edges
+
+    def test_implicit_connections(self):
+        node0a = Node(label="node 0a")
+        node0b = node0a.subnode(Node, "node 0b")
+
+        # outbound from subnode
+        node1 = Node(label="node 1", inputs={node0b: Edge()})
+
+        # inbound to subnode
+        node2a = Node(label="node 2a")
+        node2b = Node(
+            label="node 2b",
+            parent_node=node2a,
+            inputs={node0b: Edge()},
+        )
+        self.es.add(node0b, node1, node2a)
+
+        assert self.es.max_depth == 1
+
+        graph = self.es.to_networkx()
+        assert len(graph.edges) == 2
+        assert (node0b, node1) in graph.edges
+        assert (node0b, node2b) in graph.edges
+
+        with pytest.warns(ExperimentalFeatureWarning, match="implicit_flows"):
+            graph = self.es.to_networkx(add_implicit_edges=True)
+        assert len(graph.edges) == 6
+        assert (node0a, node1) in graph.edges
+        assert (node0b, node1) in graph.edges
+        assert (node0a, node2a) in graph.edges
+        assert (node0a, node2b) in graph.edges
+        assert (node0b, node2a) in graph.edges
+        assert (node0b, node2b) in graph.edges
+
+    def test_max_depth(self):
+        node0a = Node(label="node 0a")
+        node1a = node0a.subnode(
+            Node, "node 1a", inputs={node0a: Edge()}, outputs={node0a: Edge()}
+        )
+        node2a = node1a.subnode(
+            Node, "node 2a", inputs={node1a: Edge()}, outputs={node1a: Edge()}
+        )
+        node3a = node2a.subnode(
+            Node, "node 3a", inputs={node2a: Edge()}, outputs={node2a: Edge()}
+        )
+
+        node0b = Node(label="node 0b")
+        node1b = node0b.subnode(Node, "node 1b")
+        node2b = node1b.subnode(Node, "node 2b")
+        node3b = node2b.subnode(Node, "node 3b")
+
+        self.es.add(node0a, node0b)
+
+        node0a.outputs[node3b] = Edge()
+        node1a.outputs[node1b] = Edge()
+        node1a.outputs[node2b] = Edge()
+
+        g3 = self.es.to_networkx(max_depth=3)
+        assert len(g3.nodes) == 8
+        assert len(g3.edges) == 9
+        assert (node0a, node3b) in g3.edges
+        assert (node1a, node1b) in g3.edges
+        assert (node1a, node2b) in g3.edges
+        assert (node0a, node1a) in g3.edges
+        assert (node1a, node0a) in g3.edges
+        assert (node1a, node2a) in g3.edges
+        assert (node2a, node1a) in g3.edges
+        assert (node2a, node3a) in g3.edges
+        assert (node3a, node2a) in g3.edges
+
+        g2 = self.es.to_networkx(max_depth=2)
+        assert len(g2.nodes) == 6
+        assert len(g2.edges) == 7
+        assert (node0a, node2b) in g2.edges
+        assert (node1a, node1b) in g2.edges
+        assert (node1a, node2b) in g2.edges
+        assert (node0a, node1a) in g2.edges
+        assert (node1a, node0a) in g2.edges
+        assert (node1a, node2a) in g2.edges
+        assert (node2a, node1a) in g2.edges
+
+        g1 = self.es.to_networkx(max_depth=1)
+        assert len(g1.nodes) == 4
+        assert len(g1.edges) == 4
+        assert (node0a, node1b) in g1.edges
+        assert (node1a, node1b) in g1.edges
+        assert (node0a, node1a) in g1.edges
+        assert (node1a, node0a) in g1.edges
+
+        g0 = self.es.to_networkx(max_depth=0)
+        assert len(g0.nodes) == 2
+        assert list(g0.edges) == [(node0a, node0b)]
+
+        g_1 = self.es.to_networkx(max_depth=-1)
+        g_2 = self.es.to_networkx(max_depth=-2)
+        g_3 = self.es.to_networkx(max_depth=-3)
+
+        assert nx.utils.graphs_equal(g_1, g3)
+        assert nx.utils.graphs_equal(g_2, g2)
+        assert nx.utils.graphs_equal(g_3, g1)